requirements.txt
# GEOINT CLI - Dependencies
# Python 3.11+

rich>=13.7.0
requests>=2.31.0
folium>=0.15.0
geopy>=2.4.1
pyyaml>=6.0.1
dnspython>=2.4.0
astral>=3.2
# ipaddress, socket, json, logging - built-in

# Optional: face detection (OSINT face search)
# opencv-python>=4.8.0
